feat: Points are awarded based on attendance, diary writing, and comment writing logic
