+++ src/main/java/com/gdg/Todak/point/service/PointService.java
@@ -9,6 +9,7 @@ import com.gdg.Todak.point.dto.PointRequest;
 import com.gdg.Todak.point.dto.PointResponse;
 import com.gdg.Todak.point.entity.Point;
 import com.gdg.Todak.point.exception.BadRequestException;
+import com.gdg.Todak.point.exception.ConflictException;
 import com.gdg.Todak.point.exception.NotFoundException;
 import com.gdg.Todak.point.repository.PointLogRepository;
 import com.gdg.Todak.point.repository.PointRepository;
@@ -45,6 +46,16 @@ public class PointService {
     private final MemberRepository memberRepository;
     private final PointLogRepository pointLogRepository;
     private final PointLogService pointLogService;
+
+    @Transactional
+    public void createPoint(Member member) {
+        if (pointRepository.existsByMember(member)) {
+            throw new ConflictException("이미 해당 멤버의 point 객체가 존재합니다.");
+        }
+
+        Point point = Point.builder().member(member).build();
+        pointRepository.save(point);
+    }
 
     public PointResponse getPoint(String userId) {
         Member member = getMember(userId);
